* upstream/master: (267 commits)
  (maint) Update developer docs - adds parser_work
  (PUP-4917) Windows Packages Use QuietDisplayName
  (maint) Improve error message for xxx::data not found
  (PUP-5035) Eliminate extra calls to resource.to_ral
  (PUP-5035) Only add child resources for directory sources
  (maint) Test using the correct catalog
  (PUP-5278) Add test where value is found in global data binder
  (PUP-5278) Include global data binding in explain output
  (PUP-5395) Add test for regular lookup of 'lookup_options' hash.
  (PUP-5395) Fix problem with lookup_options not found in module
  (PUP-5145) Add node definition for OS X 10.11 El Capitan
  (PUP-5097) Rename domainname trusted fact to domain
  (maint) Include CentOS platform in mount util helpers
  (PUP-5432) Use node_name to validate report test
  (PUP-5441) Sanitize node's trusted info
  (PUP-5441) Add methods to ignore/restore a context key
  (PUP-5425) Fix bug with phantom node in PuppetDB
  (PUP-5433) Set solaris zone teardown to exit cleanly
  (PUP-5383) Don't restart puppetserver twice
  (PUP-5339) Update mount tests to work in AIX
  ...

+++ acceptance/tests/resource/mount/destroy.rb
@@ -2,50 +2,52 @@ test_name "should delete an entry in filesystem table and unmount it"
 
 confine :except, :platform => ['windows']
 confine :except, :platform => /osx/ # See PUP-4823
-confine :except, :platform => /solaris/
-confine :except, :platform => /aix/
+confine :except, :platform => /solaris/ # See PUP-5201
 
-fstab = '/etc/fstab'
+require 'puppet/acceptance/mount_utils'
+extend Puppet::Acceptance::MountUtils
+
 name = "pl#{rand(999999).to_i}"
 
 agents.each do |agent|
+  fs_file = filesystem_file(agent)
+
   teardown do
-    #(teardown) restore the fstab file
-    on(agent, "mv /tmp/fstab #{fstab}", :acceptable_exit_codes => (0..254))
     #(teardown) umount disk image
     on(agent, "umount /#{name}", :acceptable_exit_codes => (0..254))
     #(teardown) delete disk image
-    on(agent, "rm /tmp/#{name}", :acceptable_exit_codes => (0..254))
+    if agent['platform'] =~ /aix/
+      on(agent, "rmlv -f #{name}", :acceptable_exit_codes => (0..254))
+    else
+      on(agent, "rm /tmp/#{name}", :acceptable_exit_codes => (0..254))
+    end
     #(teardown) delete mount point
     on(agent, "rm -fr /#{name}", :acceptable_exit_codes => (0..254))
+    #(teardown) restore the fstab file
+    on(agent, "mv /tmp/fs_backup_file #{fs_file}", :acceptable_exit_codes => (0..254))
   end
 
   #------- SETUP -------#
-  step "(setup) backup fstab file"
-  on(agent, "cp #{fstab} /tmp/fstab", :acceptable_exit_codes => [0,1])
-
-  step "(setup) create disk image"
-  on(agent, "dd if=/dev/zero of=/tmp/#{name} count=10240", :acceptable_exit_codes => [0,1])
-  on(agent, "yes | mkfs -t ext3 -q /tmp/#{name}")
+  step "(setup) backup #{fs_file} file"
+  on(agent, "cp #{fs_file} /tmp/fs_backup_file", :acceptable_exit_codes => [0,1])
 
   step "(setup) create mount point"
   on(agent, "mkdir /#{name}", :acceptable_exit_codes => [0,1])
 
-  step "(setup) add entry to filesystem table"
-  on(agent, "echo '/tmp/#{name}  /#{name}  ext3  loop  0  0' >> #{fstab}")
+  step "(setup) create new filesystem to be mounted"
+  create_filesystem(agent, name)
+
+  step "(setup) add entry to the filesystem table"
+  add_entry_to_filesystem_table(agent, name)
 
   step "(setup) mount entry"
   on(agent, "mount /#{name}")
 
-  #------- TESTS -------#
   step "destroy a mount with puppet (absent)"
-  args = ['ensure=absent',
-          "device='/tmp/#{name}'"
-         ]
-  on(agent, puppet_resource('mount', "/#{name}", args))
+  on(agent, puppet_resource('mount', "/#{name}", 'ensure=absent'))
 
   step "verify entry removed from filesystem table"
-  on(agent, "cat #{fstab}") do |res|
+  on(agent, "cat #{fs_file}") do |res|
     fail_test "found the mount #{name}" if res.stdout.include? "#{name}"
   end
